LayoutTests/imported/w3c/web-platform-tests/navigation-api/ordering-and-transition/back-same-document-intercept_no-currententrychange-expected.txt


FAIL event and promise ordering for same-document navigation.back() intercepted by intercept() assert_equals: event 2 (handler run): location.hash value expected "" but got "#1"
